requirements.txt
# CIP Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Google Cloud
google-cloud-bigquery>=3.27.0
google-cloud-storage>=2.19.0
google-api-python-client>=2.159.0
google-auth>=2.37.0

# Google Sheets
gspread>=6.1.0

# Slack
slack-sdk>=3.34.0

# Firebase Auth
firebase-admin>=6.6.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0
tenacity>=9.0.0

# Dev / Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
